# docs/swagger.yaml
basePath: /
definitions:
  app.commonResponse:
    properties:
      message:
        type: string
    type: object
  app.registerPayload:
    properties:
      email:
        type: string
      organization:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - organization
    - password
    type: object
  app.resendVerificationPayload:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  app.response:
    properties:
      disposable:
        $ref: '#/definitions/app.status'
      mail:
        $ref: '#/definitions/app.status'
      mysql:
        $ref: '#/definitions/app.status'
      redis:
        $ref: '#/definitions/app.status'
      status:
        type: string
    type: object
  app.status:
    properties:
      reason:
        type: string
      status:
        type: string
    type: object
  utils.errorResponse:
    properties:
      error:
        type: string
    type: object
info:
  contact: {}
  license:
    name: Private Company License
  title: ASM Backend
  version: "1.0"
paths:
  /health:
    get:
      description: Get server health status
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.response'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/app.response'
      summary: healthCheckHandler
      tags:
      - internal
  /v1/auth/register:
    post:
      consumes:
      - application/json
      description: Register new user and organization
      parameters:
      - description: Register payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/app.registerPayload'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/app.commonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.errorResponse'
      summary: registerHandler
      tags:
      - auth
  /v1/auth/resend-email-verification:
    post:
      consumes:
      - application/json
      description: Resend verification email
      parameters:
      - description: Resend verification payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/app.resendVerificationPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.commonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.errorResponse'
        "425":
          description: Too Early
          schema:
            $ref: '#/definitions/utils.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.errorResponse'
      summary: resendEmailVerificationHandler
      tags:
      - auth
swagger: "2.0"